+++ src/index.tsx
@@ -2,13 +2,10 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import { store } from './store'
 import { BrowserRouter } from 'react-router-dom'
-
 import './firebase'
 import { Provider } from 'react-redux'
-import { store } from './store/store'
 import App from './App'
 import './firebase'
-
 
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement)
 root.render(
