+++ src/store/hooks/useAppHook.ts
@@ -1,5 +1,5 @@
 import { useDispatch, useSelector, TypedUseSelectorHook } from 'react-redux'
-import { AppDispatch, RootState } from '../store'
+import { AppDispatch, RootState } from '../../store'
 
 export const useAppDispatch = () => useDispatch<AppDispatch>()
 export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector
